pyproject.toml
[project]
name = "patchflow-run-checker"
version = "0.1.0"
description = "A script that checks Supabase for pending private patchflow runs and triggers the patchflow command"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "supabase>=2.15.0",
]
